Former-commit-id: 7573bc1f4994892659d7916ac59e8359cdc4faca

+++ app/src/main/java/com/imcys/bilibilias/base/utils/DialogUtils.kt
@@ -453,7 +453,7 @@ object DialogUtils {
 
     private fun addThirdPartyData(
         bvid: String,
-        aid: Int,
+        aid: Long,
         mid: Long,
         name: String,
         copyright: Int,
