+++ app/src/main/java/com/imcys/bilibilias/home/ui/viewmodel/AsVideoViewModel.kt
@@ -366,7 +366,7 @@ class AsVideoViewModel : ViewModel() {
      * 加载用户收藏夹
      */
     @SuppressLint("NotifyDataSetChanged")
-    fun loadCollectionView(view: View, avid: Int) {
+    fun loadCollectionView(view: View, avid: Long) {
         val context = view.context
         (context as AsVideoActivity).binding.apply {
             viewModelScope.launch(Dispatchers.IO) {
@@ -417,7 +417,7 @@ class AsVideoViewModel : ViewModel() {
      * 设置收藏夹的ID列表
      * @param selects MutableList<Long>
      */
-    private fun setCollection(context: AsVideoActivity, selects: MutableList<Long>, avid: Int) {
+    private fun setCollection(context: AsVideoActivity, selects: MutableList<Long>, avid: Long) {
         var addMediaIds = ""
         selects.forEachIndexed { index, l ->
             if (index == selects.size) {
@@ -432,7 +432,7 @@ class AsVideoViewModel : ViewModel() {
      * 新增收藏夹内容
      * @param addMediaIds String
      */
-    private fun addCollection(context: AsVideoActivity, addMediaIds: String, avid: Int) {
+    private fun addCollection(context: AsVideoActivity, addMediaIds: String, avid: Long) {
         viewModelScope.launch(Dispatchers.Default) {
             val collectionResultBean =
                 KtHttpUtils.addHeader(COOKIE, context.asUser.cookie)
